Added password confirmation to create account

+++ job_board/frontend/src/components/Auth/RegisterForm.tsx
@@ -2,12 +2,14 @@ import { useForm, Controller } from 'react-hook-form';
 import { FormControl, FormControlLabel, RadioGroup, Radio, Typography, TextField, Button, Stack } from '@mui/material';
 import { useState } from 'react';
 import { userRegister } from '../../services/auth';
+import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
 
 type FormData = {
     name: string,
     role: string,
     email: string;
     password: string;
+    confirmPassword: string;
 };
 
 function RegisterForm() {
@@ -19,6 +21,7 @@ function RegisterForm() {
             name: '',
             email: '',
             password: '',
+            confirmPassword: '',
             role: '',
         },
     });
@@ -26,12 +29,16 @@ function RegisterForm() {
     const onSubmit = async (data: FormData) => {
         try {
             setIsLoading(true);
+            if (data.password !== data.confirmPassword) {
+                setErrorMessage('Passwords do not match.');
+                return
+            }
             const response = await userRegister(data.role, data.name, data.email, data.password);
             localStorage.setItem('token', response.token);
         } catch (error) {
-            const errorMessage = error instanceof Error 
-            ? `${error.message}. Please, check your credentials.` 
-            : 'An error occurred';
+            const errorMessage = error instanceof Error
+                ? `${error.message}. Please, check your credentials.`
+                : 'An error occurred';
             setErrorMessage(errorMessage);
         } finally {
             setIsLoading(false);
@@ -40,10 +47,10 @@ function RegisterForm() {
 
     return (
         <>
-            <form onSubmit={handleSubmit(onSubmit)}>
+            <form onSubmit={handleSubmit(onSubmit)} className='mb-4'>
                 <FormControl>
                     <Stack spacing={2} className='w-72 lg:w-96'>
-                    <Controller
+                        <Controller
                             name="name"
                             control={control}
                             rules={{ required: 'Name is required' }}
@@ -86,6 +93,20 @@ function RegisterForm() {
                             )}
                         />
                         <Controller
+                            name="confirmPassword"
+                            control={control}
+                            rules={{ required: 'Password confirmation is required' }}
+                            render={({ field, fieldState }) => (
+                                <TextField
+                                    label="Password"
+                                    type="password"
+                                    error={!!fieldState.error}
+                                    helperText={fieldState.error ? fieldState.error.message : ''}
+                                    {...field}
+                                />
+                            )}
+                        />
+                        <Controller
                             name="role"
                             control={control}
                             rules={{ required: 'Type is required' }}
@@ -97,8 +118,8 @@ function RegisterForm() {
                                         value={field.value}
                                         onChange={(e) => field.onChange(e.target.value)}
                                     >
-                                        <FormControlLabel value="person" control={<Radio />} label="I'm looking for a job."  />
-                                        <FormControlLabel value="company" control={<Radio />} label="I'm looking for talent."  />
+                                        <FormControlLabel value="person" control={<Radio />} label="I'm looking for a job." />
+                                        <FormControlLabel value="company" control={<Radio />} label="I'm looking for talent." />
                                     </RadioGroup>
                                     {fieldState.error && (
                                         <Typography style={{ color: 'red' }} variant="caption" color="textSecondary">{fieldState.error.message}</Typography>
@@ -106,7 +127,9 @@ function RegisterForm() {
                                 </>
                             )}
                         />
-                        <span>{errorMessage}</span>
+                        {errorMessage && (
+                            <span className="text-primary mt-2 text-center"><ErrorOutlineIcon /> {errorMessage}</span>
+                        )}
                         <Button type="submit" variant="contained" color="primary" disabled={isLoading}>
                             {isLoading ? 'Creating account...' : 'Create my account'}
                         </Button>
